Restructure Tile movement
Refactor the updated player movement to be for all Tiles
